requirements.txt
paho_mqtt==2.1.0
requests